requirements.txt
streamlit
fastapi
python-dotenv
sqlalchemy
psycopg2
torch
sklearn
transformers
peft
joblib
pandas
numpy
qdrant-client
uvicorn
python-multipart
beautifulsoup4
sentence-transformers
imblearn
streamlit-modal
st-star-rating
